OBElementTable::GetAtomicNum: Changed unsigned short int& to int&. More consistent with other ints since have to use the exact int type when calling. Now used in mdlformat.cpp.

+++ src/data.h
@@ -138,7 +138,7 @@ public:
     //! Returns the atomic number matching the element symbol passed
     //! or 0 if not defined. For 'D' or 'T' hydrogen isotopes, will return
     //! a value in the second argument
-    int   GetAtomicNum(const char *, unsigned short int &iso);
+    int   GetAtomicNum(const char *, int &iso);
     //! Returns the element symbol matching the atomic number passed
     char *GetSymbol(int);
     //! Returns the van der Waals radius for this atomic number
